fixing major issue with thread preemption
the PIC end-of-interrupt wasn't being sent when a thread is
preempted in the PIT handler because if the next thread is
being launched for the first time, `launch_thread` is called
followed by the thread-specific code instead of returning to
the IRQ handler. Now a global flag `g_need_reschedule` is set
and checked after the interrupt is handled. This needs cleaned up.